girlscout: move/rename various files
